Removed resiliency constructs
