+++ crates/wasi-http/src/host/default_impl.rs
@@ -1,5 +1,4 @@
 use std::fmt::Display;
-use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::{Context, Result};
@@ -21,10 +20,6 @@ use wasmtime_wasi::TrappableError;
 use wasmtime_wasi_http::WasiHttpCtx;
 use wasmtime_wasi_http::p3::bindings::http::types::ErrorCode;
 use wasmtime_wasi_http::p3::{self, RequestOptions, WasiHttpCtxView};
-
-use super::circuit_breaker::{BreakerConfig, BucketRegistry};
-use super::outbound::{self, ResilienceConfig};
-use super::retry::RetryPolicy;
 
 pub type HttpResult<T> = Result<T, HttpError>;
 pub type HttpError = TrappableError<ErrorCode>;
@@ -49,26 +44,6 @@ pub struct ConnectOptions {
     pub addr: String,
     #[env(from = "HTTP_CONNECT_TIMEOUT_SECS", default = "10")]
     pub connect_timeout_secs: u64,
-    #[env(from = "HTTP_OUTBOUND_RESILIENCE", default = "false")]
-    pub outbound_resilience: bool,
-    #[env(from = "HTTP_RESPONSE_TIMEOUT_MS", default = "0")]
-    pub response_timeout_ms: u64,
-    #[env(from = "HTTP_RETRY_MAX", default = "2")]
-    pub retry_max: u8,
-    #[env(from = "HTTP_RETRY_BASE_DELAY_MS", default = "100")]
-    pub retry_base_delay_ms: u64,
-    #[env(from = "HTTP_RETRY_CAP_DELAY_MS", default = "1000")]
-    pub retry_cap_delay_ms: u64,
-    #[env(from = "HTTP_CB_TRIP_THRESHOLD", default = "10")]
-    pub cb_trip_threshold: u32,
-    #[env(from = "HTTP_CB_RECOVERY_THRESHOLD", default = "5")]
-    pub cb_recovery_threshold: u32,
-    #[env(from = "HTTP_CB_RESET_PERIOD_MS", default = "10000")]
-    pub cb_reset_period_ms: u64,
-    #[env(from = "HTTP_CB_FAULT_WINDOW_MS", default = "30000")]
-    pub cb_fault_window_ms: u64,
-    #[env(from = "HTTP_CB_BUCKETS", default = "")]
-    pub cb_buckets: String,
 }
 
 impl omnia::FromEnv for ConnectOptions {
@@ -82,8 +57,6 @@ impl omnia::FromEnv for ConnectOptions {
 struct HttpHooks {
     client: reqwest::Client,
     connect_timeout: Duration,
-    default_timeout: Duration,
-    resilience: Option<ResilienceConfig>,
 }
 
 /// Default implementation for `wasi:http`.
@@ -110,7 +83,6 @@ impl Backend for HttpDefault {
     #[instrument]
     async fn connect_with(options: Self::ConnectOptions) -> Result<Self> {
         let connect_timeout = Duration::from_secs(options.connect_timeout_secs);
-        let default_timeout = Duration::from_millis(options.response_timeout_ms);
 
         let builder = reqwest::Client::builder().connect_timeout(connect_timeout);
 
@@ -118,46 +90,10 @@ impl Backend for HttpDefault {
         let builder = builder.no_proxy();
 
         let client = builder.build().context("building HTTP client")?;
-
-        let resilience = if options.outbound_resilience {
-            let breaker_config = BreakerConfig {
-                trip_threshold: options.cb_trip_threshold,
-                recovery_threshold: options.cb_recovery_threshold,
-                reset_period: Duration::from_millis(options.cb_reset_period_ms),
-                fault_window: Duration::from_millis(options.cb_fault_window_ms),
-            };
-
-            let bucket_names = options
-                .cb_buckets
-                .split(',')
-                .map(str::trim)
-                .filter(|s| !s.is_empty())
-                .map(String::from);
-
-            let registry = BucketRegistry::new(bucket_names, &breaker_config)
-                .context("building circuit breaker registry")?;
-
-            let retry_policy = RetryPolicy {
-                base_delay_ms: options.retry_base_delay_ms,
-                cap_delay_ms: options.retry_cap_delay_ms,
-            };
-            retry_policy.validate().context("validating retry policy")?;
-
-            Some(ResilienceConfig {
-                retry_max: options.retry_max,
-                retry_policy,
-                registry: Arc::new(registry),
-            })
-        } else {
-            None
-        };
-
         Ok(Self {
             hooks: HttpHooks {
                 client,
                 connect_timeout,
-                default_timeout,
-                resilience,
             },
             ctx: WasiHttpCtx::default(),
         })
@@ -175,14 +111,15 @@ impl p3::WasiHttpHooks for HttpHooks {
     > {
         let shared_client = self.client.clone();
         let connect_timeout = self.connect_timeout;
-        let default_timeout = self.default_timeout;
-        let resilience = self.resilience.clone();
 
         Box::new(async move {
             let (mut parts, body) = request.into_parts();
 
-            // Use a one-off client when a client certificate is required, otherwise
-            // reuse the shared client for connection pooling.
+            // remove "Host" headers (`reqwest` adds its own)
+            parts.headers.remove(HOST);
+
+            // use a one-off client when a client certificate is required, otherwise
+            // reuse the shared client for connection pooling
             let client = if let Some(encoded_cert) = parts.headers.remove("Client-Cert") {
                 tracing::debug!("using client certificate");
                 let encoded = encoded_cert.to_str().map_err(internal_err)?;
@@ -194,23 +131,30 @@ impl p3::WasiHttpHooks for HttpHooks {
                 #[cfg(test)]
                 let builder = builder.no_proxy();
 
-                builder.build().map_err(outbound::reqwest_to_error_code)?
+                builder.build().map_err(reqwest_err)?
             } else {
                 shared_client
             };
 
-            let request = Request::from_parts(parts, body);
-            let resp = outbound::send(&client, request, resilience.as_ref(), default_timeout)
-                .await
-                .map_err(HttpError::from)?;
+            let collected = body.collect().await.map_err(internal_err)?;
 
-            // Process response
+            // make request
+            let url = parts.uri.to_string();
+            let resp = client
+                .request(parts.method, &url)
+                .headers(parts.headers)
+                .body(collected.to_bytes())
+                .send()
+                .await
+                .map_err(reqwest_err)?;
+
+            // process response
             let converted: Response<reqwest::Body> = resp.into();
             let (parts, body) = converted.into_parts();
-            let body = body.map_err(outbound::reqwest_to_error_code).boxed_unsync();
+            let body = body.map_err(reqwest_err).boxed_unsync();
             let mut response = Response::from_parts(parts, body);
 
-            // Remove forbidden headers (disallowed by `wasmtime-wasi-http`)
+            // remove forbidden headers (disallowed by `wasmtime-wasi-http`)
             let headers = response.headers_mut();
             for header in &FORBIDDEN_HEADERS {
                 headers.remove(header);
@@ -223,6 +167,19 @@ impl p3::WasiHttpHooks for HttpHooks {
 
 fn internal_err(e: impl Display) -> ErrorCode {
     ErrorCode::InternalError(Some(e.to_string()))
+}
+
+#[allow(clippy::needless_pass_by_value)]
+fn reqwest_err(e: reqwest::Error) -> ErrorCode {
+    if e.is_timeout() {
+        ErrorCode::ConnectionTimeout
+    } else if e.is_connect() {
+        ErrorCode::ConnectionRefused
+    } else if e.is_request() {
+        ErrorCode::HttpRequestUriInvalid
+    } else {
+        internal_err(e)
+    }
 }
 
 #[cfg(test)]
@@ -242,16 +199,6 @@ mod tests {
         let options = ConnectOptions {
             addr: String::new(),
             connect_timeout_secs: 10,
-            outbound_resilience: false,
-            response_timeout_ms: 0,
-            retry_max: 2,
-            retry_base_delay_ms: 100,
-            retry_cap_delay_ms: 1000,
-            cb_trip_threshold: 10,
-            cb_recovery_threshold: 5,
-            cb_reset_period_ms: 10_000,
-            cb_fault_window_ms: 30_000,
-            cb_buckets: String::new(),
         };
         HttpDefault::connect_with(options).await.unwrap()
     }
@@ -425,96 +372,5 @@ mod tests {
             let boxed = self.hooks.send_request(request, None, Box::new(async { Ok(()) }));
             Pin::from(boxed).await
         }
-    }
-
-    // --- Integration tests for resilience glue ---
-    //
-    // Resilience logic (retry, circuit breaker, timeout, header handling) is
-    // thoroughly tested in `outbound::tests`. Tests here verify only the
-    // integration surface that `default_impl` adds on top: the `WasiHttpHooks`
-    // trait wiring, client-cert branching, forbidden-header stripping on the
-    // response, and the resilience-disabled opt-out path.
-
-    async fn resilience_client(timeout_ms: u64, retry_max: u8, cb_threshold: u32) -> HttpDefault {
-        let options = ConnectOptions {
-            addr: String::new(),
-            connect_timeout_secs: 10,
-            outbound_resilience: true,
-            response_timeout_ms: timeout_ms,
-            retry_max,
-            retry_base_delay_ms: 10,
-            retry_cap_delay_ms: 50,
-            cb_trip_threshold: cb_threshold,
-            cb_recovery_threshold: 2,
-            cb_reset_period_ms: 100,
-            cb_fault_window_ms: 30_000,
-            cb_buckets: String::new(),
-        };
-        HttpDefault::connect_with(options).await.unwrap()
-    }
-
-    #[tokio::test]
-    async fn send_request_preserves_existing_behavior() {
-        let server = MockServer::start().await;
-        Mock::given(method("GET"))
-            .respond_with(ResponseTemplate::new(200).set_body_string("preserved"))
-            .mount(&server)
-            .await;
-
-        let mut client = test_client().await;
-        let request = Request::get(server.uri())
-            .body(Empty::new().map_err(internal_err).boxed_unsync())
-            .unwrap();
-
-        let result = client.handle(request).await;
-        assert!(result.is_ok());
-        let (resp, _) = result.unwrap();
-        assert_eq!(resp.status(), StatusCode::OK);
-        let body = resp.into_body().collect().await.unwrap().to_bytes();
-        assert_eq!(body, Bytes::from("preserved"));
-    }
-
-    #[tokio::test]
-    async fn send_request_client_cert_with_resilience() {
-        let server = MockServer::start().await;
-        Mock::given(method("GET")).respond_with(ResponseTemplate::new(200)).mount(&server).await;
-
-        let mut client = resilience_client(5000, 2, 10).await;
-
-        let request = Request::get(server.uri())
-            .header("Client-Cert", "not-valid-base64!!!")
-            .body(Empty::new().map_err(internal_err).boxed_unsync())
-            .unwrap();
-
-        let result = client.handle(request).await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn send_request_resilience_disabled() {
-        let server = MockServer::start().await;
-        Mock::given(method("GET")).respond_with(ResponseTemplate::new(503)).mount(&server).await;
-
-        let mut client = test_client().await;
-
-        let request = Request::get(server.uri())
-            .body(Empty::new().map_err(internal_err).boxed_unsync())
-            .unwrap();
-
-        let (resp, _fut) = client.handle(request).await.unwrap();
-        assert_eq!(resp.status().as_u16(), 503);
-
-        let received = server.received_requests().await.unwrap();
-        assert_eq!(received.len(), 1, "resilience off: GET should not retry");
-
-        for _ in 0..5 {
-            let request = Request::get(server.uri())
-                .body(Empty::new().map_err(internal_err).boxed_unsync())
-                .unwrap();
-            let _ = client.handle(request).await;
-        }
-
-        let received = server.received_requests().await.unwrap();
-        assert_eq!(received.len(), 6, "all 6 requests should reach server without breaker");
     }
 }
